backend/requirements.txt
Django
djangorestframework
djangorestframework-simplejwt
django-cors-headers
psycopg2-binary
gunicorn
whitenoise
dj-database-url
python-dotenv
requests
